moved all methods to UserAction.cs, freeing up some code
